+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### 0.0.3 (Oct 20 2024)
+
+* `m:Thumbtack.ImageUpload`:
+  * `:format` option added
+  * callback API changed
+* `:resize` image transformation now supports rectangles
+
+
 ### 0.0.2 (Oct 19 2024)
 
 * `m:Thumbtack.ImageUpload` callback API changed